+++ src/test/java/store/sonyk9919/api/domain/island/service/IslandLevelServiceConcurrencyTest.java
@@ -6,14 +6,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import store.sonyk9919.api.domain.auth.service.KakaoProvider;
-import store.sonyk9919.api.global.file.resolver.FilePathResolver;
 import store.sonyk9919.api.support.IntegrationTestSupport;
 
 import java.util.concurrent.CountDownLatch;
@@ -22,18 +14,7 @@ import java.util.concurrent.Executors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@Testcontainers
 class IslandLevelServiceConcurrencyTest extends IntegrationTestSupport {
-
-    @Container
-    private static final GenericContainer<?> redis =
-            new GenericContainer<>("redis:7-alpine").withExposedPorts(6379);
-
-    @DynamicPropertySource
-    static void setRedisProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.data.redis.host", redis::getHost);
-        registry.add("spring.data.redis.port", () -> redis.getMappedPort(6379));
-    }
 
     @Autowired
     private IslandLevelService islandLevelService;
