+++ src/test/java/store/sonyk9919/api/support/IntegrationTestSupport.java
@@ -2,11 +2,27 @@ package store.sonyk9919.api.support;
 
 import org.springframework.amqp.rabbit.connection.ConnectionFactory;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 import store.sonyk9919.api.domain.auth.service.KakaoProvider;
 
 @SpringBootTest
+@Testcontainers
 public abstract class IntegrationTestSupport {
+
+    @Container
+    private static final GenericContainer<?> redis =
+            new GenericContainer<>("redis:7-alpine").withExposedPorts(6379);
+
+    @DynamicPropertySource
+    static void setRedisProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.data.redis.host", redis::getHost);
+        registry.add("spring.data.redis.port", () -> redis.getMappedPort(6379));
+    }
 
     @MockitoBean protected ConnectionFactory connectionFactory;
